Make ‘make check’ actually *test* and not just *run*.
Specifically, it now compares the output log of all the example programs
with a file of expected output, after filtering out some predictable
forms of nondeterminism.  This is crude, but it’s better than nothing.

The filter can handle changes in values printed (e.g. execution
timings) but not changes in output order, so examples/shuffle.c has to
be changed to use a fixed shuffle seed.  Some day maybe we will get a
more sophisticated filter (and more easily parseable test output) and
this can be reverted.

Also fix an oversight in the Makefile that meant we weren’t applying
any warning options, and fix the warnings that popped up when that was
corrected.  (Notably, we have had to disable ITEST_USE_SETJMP mode in
order to suppress some ‘variable might be clobbered by longjmp’ warnings.
This is an actual bug in Infinitestimal that will be corrected shortly.)