Remove ITEST_MAIN_BEGIN and ITEST_MAIN_END.
The canonical test runner main() now begins with

    itest_init();
    itest_parse_options(argc, argv);

and ends with

    return itest_print_report();

This avoids hiding control flow, and makes it clearer what the
difference is between a standalone test runner program and a test
suite embedded in a larger program — namely, the call to
itest_parse_options.

Also teach ‘make check’ to monitor for changes in exit codes.