Move body of RUN_TEST*, PASS, FAIL, SKIP to itest.c.
In addition to slimming down itest.h, this should make it safe to
re-enable longjmp, as the calls to both setjmp and longjmp are
isolated from any variables they might “clobber” (that bug shoulda
been fixed decades ago, feh).  On the other hand, it means
ITEST_RUN_TESTp has to be removed…but I have a better idea, so that’s
ok.